fix warnning

+++ src/icmp_protocol.hpp
@@ -37,6 +37,6 @@ typedef struct _ECHORESPONSE {
 } ECHORESPONSE,*PECHORESPONSE;
 
 
-int sendEchoRequest(SOCKET sock , sockaddr_in detIP);
-int recvEchoReQuest(SOCKET sock,ECHORESPONSE * sponse, sockaddr_in *dstIP);
+ssize_t sendEchoRequest(SOCKET sock , sockaddr_in detIP);
+ssize_t recvEchoReQuest(SOCKET sock,ECHORESPONSE * sponse, sockaddr_in *dstIP);
 u_short checksum(u_short *buffer, int len);
